requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
PyYAML>=6.0.1
aiohttp>=3.8.0
python-dateutil>=2.8.2
